pyproject.toml
[project]
name = "a2a_experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.24",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
